% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/turn-package.R
\docType{package}
\name{turn-package}
\alias{turn}
\alias{turn-package}
\title{turn: What the Package Does (One Line, Title Case)}
\description{
Convert files in multiple format
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/datasketch/turn}
  \item Report bugs at \url{https://github.com/datasketch/turn/issues}
}

}
\author{
\strong{Maintainer}: Juan Pablo Marin Diaz \email{jpmarindiaz@gmail.com}

}
\keyword{internal}
